requirements.txt
anthropic
python-dotenv
pyyaml
rumps
pywebview
